requirements.txt
runpod==1.6.2
torch==2.1.2
torchvision==0.16.2
diffusers==0.27.2
transformers==4.38.2
accelerate==0.27.2
xformers==0.0.25
safetensors==0.4.2
opencv-python-headless==4.9.0.80
pillow==10.2.0
requests==2.31.0